[embedlite-components] Remove no longer needed zoomToRect calls. Fixes JB#27546.
The code was added originally in PR #35 to fix viewport update problems
in cases where the keyboard was open. The original issue is no longer
reproducible without the patch, but the code added by it causes serious
problems on twitter.com. It triggers additional scrolls after the
viewport size is changed due to page custom HTML dialogs being shown.