[embedlite-components] Add support for site specific UA override. Contributes to JB#34904
This makes possible to use general.useragent.override.* overrides
with simple replace syntax given that complex override is not applied.

Instantiation of the UserAgent is deferred a bit so that we do not
end up initializing UserAgentOverrides before general.useragent.override
is applied. As UserAgent was initialized on that application startup,
the simple hash separated user agent fixing did not work. This happened
because nsIHttpProtocolHandler was not yet aware of the override.
Thus, causing UserAgentOverrides to fail to replace ua.