[embedlite-components] Fixes out of bounds error of session history. Fixes JB#36736
When a tab was used longer time so that back history exceeded
"browser.sessionhistory.max_entries" preference value, we hit out of
bounds error and a wrong url was shown in the toolbar.

As "embedui:addhistory" needs to receive sorted list we can
drop "extra items" from the beginning of the list and adjust
session history index accordingly.