[embedlite-components] Use promptForSaveToFileAsync. Contributes to JB#49723
The promptForSaveToFile() method, used to determine the path to save
downloaded files to (usually the Downloads directory) has been removed
in favour of promptForSaveToFileAsync().

This change updates the code to use the async version.

The previous version had code for prompting the user with a file
dialogue box, but it didn't seem operational, so this has been removed.