[embedlite-components] Ensure touch events also fire mouse clicks. Fixes JB#45659
Gesture single taps were previously configured to send out mouse events
only if they were preceded by a touchstart event, and only if
preventDefault() wasn't applied to the event.

Other browsers send out the event independent of this. The difference
manifests itself if stopPropagation() is applied to the touch event,
which supresses the mouse events on our browser, but not on others.

For example, this meant that the input field of DuckDuckGo couldn't be
focussed, and also prevented the Google Maps touch controls from
working.

This change alters things so that the mouse event is sent out even if
single tap isn't preceeded by a touch event.