[embedlite-components] Ensure restored history entries have a triggeringPrincipal. Fixes JB#51572
When the history entries are restored when the browser starts up, or on
a tab change, the triggeringPrincipal must be set on them, otherwise
DocShell won't allow them to be used with the GoBack()/GoForwards()
methods.

This change adds a NullPrincipal(), which is considered a valid value
(i.e. not a null pointer).