[embedlite-components] Print channel info to the logs. Contributes to JB#50424
Provides information about channels when opening a website. When active
the following info is output to the log:

1. Request type and URL.
2. Response status.
3. Request and response headers.
4. A dump of the returned content if it's textual.

The aim is to provide something a little similar to the Network tab of
the developer tools in Firefox.

The mode is activated by including the string "network" in the
EMBED_CONSOLE environment variable. This can be combined with others,
for example:

EMBED_CONSOLE="stacktrace,network" sailfish-browser