[embedlite-components] Handle form actions inside the captive portal. JB#54920
Any form actions the user triggers while in the captive portal will be
handled inside the captive portal, rather than breaking out into the
browser, independent of whether they are in the same origin as the page
or not.

However, any future clicks will break out if they don't match the
origin of the form page, rather than of any subsequent pages loaded.

This change is intended to ensure form actions inside the captive portal
still work as expected; often needed when logging in to a wifi provider.