[embedlite-components] Replace uses of URI only permissions API. JB#56082 OMP#JOLLA-485
The DOMPopupBlocked event supplies a URI and not a directly derived
principal so it is still URI based just with the extra step of
constructing a principal from the URI to passed to getAllForPrincipal.

In ContentPermissionPrompt.js a principal is available from the source
but this gets fed through C++ models in sailfish-components-webview and
the limited representation means reducing it to an origin URI and the
constructing a new content principal on use.