[embedlite-components] Update to new nsILoginManagerPrompter interface. JB#55760, OMP#JOLLA-418
Updates nsLoginManagerPrompter.js to use the updated
nsILoginManagerPrompter interface of ESR78. The main change is that
rather than passing the browser during init, it instead passes it each
time a username/password is to be requested from the user.

The intention for the change is that the window can be shown in the
correct tab, but practically speaking it doesn't affect behaviour for
sailfish-browser.