[embedlite-components] Pass private browsing status when requesting permissions. JB#55925 OMP#JOLLA-445
When in private browsing mode, it shouldn't be possible to store
permissions, either permanently or for the session, since this would
leak website location information to normal mode.

As well as blocking storage in private mode, this change also passes the
status of private browsing to the UI so that it can disable the option
to save permission info when in private browsing mode.

This relies on the small addition of
LocationPopupInterface.privateBrowsing to the brower popups API.

Co-authored-by: Raine Mäkeläinen <raine.makelainen@jolla.com>