[embedlite-components] Reinstate search in text selection toolbar. Fixes JB#56187 OMP#JOLLA-505
The nsISearchService.currentEngine parameter is no longer available in
ESR78. However, we were using it to determine whether or not to display
the search button in the text selection toolbar. Since it returned
"undefined", the button was always hidden.

As it happens we already use nsISearchService.defaultEngine to
specify the search engine, so this changes the toolbar button to use
this instead.