[embedlite-components] Fix TLS certificate error page. JB#56093 OMP#JOLLA-491
Fix AboutRedirector to use the up to date newChannel() function
signature and use Services.io.newChannelFromURIWithLoadInfo() instead of
newChannel() the call to which was failing because it required
additional arguments.

Then fix EmbedLiteErrorPageHandler to remove the use of the
nsIBadCertListener2 interface which no longer exists by porting the
changes from https://phabricator.services.mozilla.com/D21791.