[embedlite-components] Populate the technical details of the certificate page. JB#56370 OMP#JOLLA-541
The description of the error is no longer encoded in the about:certerror
URL and needs to be constructed in the page itself.

Also fix the caret icon so it's apparent when the collapsed headings have
been expanded as with neither that nor content it wasn't apparent when
clicking worked.