[embedlite-components] Avoid failing login if saving credentials fails. JB#63276
At the moment a saved login info is not passed properly, hence if user
chooses 'remember' for credentials when loggin again, the related
code will try to add the credentials, failing on already existing.
And here the exception made the whole login failing.

For robustness, let's keep going with the login even if there are
problems on storing the credentials for later use.